+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## 2.20 [unreleased]
 
+### Features
+
+- Added signature to InfluxDBMapper.query() with params final Query query, final Class<T> clazz, final String measurementName to leverage InfluxDBResultMapper.toPojo method with identical signature.
+
+### Improvements
+
+- Test: Added test for new InfluxDBMapper.query() signature, as well as test for existing InfluxDBMapper.query(Class clazz) signature (previously only InfluxDBMapper.query(Query query, Class clazz) was tested).
+
 ## 2.19 [2020-05-18]
 
 ## 2.18 [2020-04-17]